Small cleaning up.
M    src/Makefile
M    src/Geometry.hpp
M    src/OptimizeMatrix.cpp
M    testing/SparseMatrix/SparseMatrixTest.cpp
M    testing/SparseMatrix/Makefile
M    testing/CG/CGTest.cpp
M    testing/CG/Makefile


git-svn-id: svn+ssh://software.sandia.gov/svn/hpcg/trunk@32 5e06a57e-6b39-4354-92a0-249a7d0c5410

+++ src/Geometry.hpp
@@ -11,7 +11,7 @@
 #ifndef GEOMETRY_HPP
 #define GEOMETRY_HPP
 
-typedef int local_int_t;
+typedef long long local_int_t;
 typedef long long global_int_t;
 
 // This is a data structure to contain all processor navigation information
